+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Rainbow CSV for Visual Studio Code Change Log
 
+## 3.9.0
+* Adjust state transition logic and improve UX for some edge cases.
+
 ## 3.8.0
 * UI improvement: Better "CSVLint" status button by [@kostasx](https://github.com/kostasx).
 * Add "Rainbow ON" conditional button.